@@ -1,8 +1,10 @@
 package com.code.to.learn.api.util;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
+import org.springframework.web.server.ResponseStatusException;
 
 @Component
 public class UsernameGetter {
@@ -12,6 +14,6 @@ public class UsernameGetter {
         if (principal instanceof UserDetails) {
             return ((UserDetails) principal).getUsername();
         }
-        return (String) principal;
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Not authorized");
     }
 }
